+++ packages/cli/src/lib/commands/collect/utils/user-flow/load-flow.test.ts
@@ -2,24 +2,20 @@ import { join } from 'path';
 import { DEFAULT_COLLECT_UF_PATH } from '../../options/ufPath.constant';
 import { loadFlow } from './load-flow';
 import {
+  INITIATED_PRJ_CFG,
+  VALIDE_EXAMPLE_USERFLOW_CONTENT,
+  VALIDE_EXAMPLE_USERFLOW_NAME,
+  WRONG_EXT_USERFLOW_CONTENT,
+  WRONG_EXT_USERFLOW_NAME,
+  WRONG_MOD_EXPORT_USERFLOW_CONTENT,
+  WRONG_MOD_EXPORT_USERFLOW_NAME
+} from 'test-data';
+import {
   UserFlowCliProject,
   UserFlowCliProjectFactory,
   UserFlowProjectConfig
 } from '@push-based/user-flow-cli-testing';
-import { INITIATED_PRJ_CFG } from '../../../../../../tests/fixtures/sandbox/initiated';
-import {
-  VALIDE_EXAMPLE_USERFLOW_CONTENT,
-  VALIDE_EXAMPLE_USERFLOW_NAME
-} from '../../../../../../tests/fixtures/user-flows/valide.example.uf';
 import { DEFAULT_RC_NAME } from '../../../../constants';
-import {
-  WRONG_EXT_USERFLOW_CONTENT,
-  WRONG_EXT_USERFLOW_NAME
-} from '../../../../../../tests/fixtures/user-flows/wrong-ext.example.uf';
-import {
-  WRONG_MOD_EXPORT_USERFLOW_CONTENT,
-  WRONG_MOD_EXPORT_USERFLOW_NAME
-} from '../../../../../../tests/fixtures/user-flows/wrong-mod-export.example.uf';
 import { DEFAULT_PERSIST_OUT_PATH } from '../../options/outPath.constant';
 
 const rcFile = INITIATED_PRJ_CFG?.rcFile;
