+++ packages/cli/src/lib/index.ts
@@ -1,4 +1,9 @@
+export { SETUP_CONFIRM_MESSAGE } from './commands/init/constants';
 export { PROMPT_COLLECT_URL } from './commands/collect/options/url.constant';
+export {  } from './commands/collect/options/ufPath';
+export { PROMPT_COLLECT_UF_PATH } from './commands/collect/options/ufPath.constant';
+export { PROMPT_PERSIST_OUT_PATH } from './commands/collect/options/outPath.constant';
+export { PROMPT_PERSIST_FORMAT } from './commands/collect/options/format.constant';
 export { ERROR_PERSIST_FORMAT_WRONG } from './commands/collect/options/format.constant';
 export { CollectArgvOptions } from './commands/collect/options/types';
 export { getInitCommandOptionsFromArgv } from './commands/init/utils';
@@ -13,7 +18,7 @@ export { MeasureModes } from './commands/collect/utils/replay/types';
 export { RcJson } from './types';
 export { CLI_MODES, CI_PROPERTY, CLI_MODE_PROPERTY } from './global/cli-mode';
 export { LhConfigJson } from './hacky-things/lighthouse';
-export { getEnvPreset, SANDBOX_PRESET,CI_PRESET, DEFAULT_PRESET } from './pre-set';
+export { getEnvPreset, SANDBOX_PRESET, CI_PRESET, DEFAULT_PRESET } from './pre-set';
 export { InitCommandArgv } from './commands/init/options/types';
 export { GlobalOptionsArgv } from './global/options/types';
 export { CollectCommandArgv } from './commands/collect/options/types';
