--- packages/sandbox/src/lib/sandbox.ts
@@ -1,3 +1,0 @@
-export function sandbox(): string {
-  return 'sandbox';
-}
