+++ packages/test-data/src/lib/budget/index.ts
@@ -1,0 +1,1 @@
+export * from './lh-navigation-budget';
