+++ packages/test-data/src/lib/config/index.ts
@@ -1,0 +1,1 @@
+export * from './lh-config';
