+++ packages/test-data/src/lib/sandboxes/cli-bin-path.ts
