+++ packages/test-data/src/lib/sandboxes/empty/index.ts
@@ -1,0 +1,1 @@
+export * from './cfg';
