+++ packages/test-data/src/lib/sandboxes/index.ts
@@ -1,0 +1,5 @@
+export * from './cli-bin-path';
+export * from './empty';
+export * from './initialized';
+export * from './remote';
+export * from './static';
