+++ packages/test-data/src/lib/sandboxes/initialized/index.ts
@@ -1,0 +1,2 @@
+export * from './cfg';
+export * from './flow1.uf';
