+++ packages/test-data/src/lib/sandboxes/remote/cfg.ts
@@ -1,8 +1,8 @@
-import { UserFlowProjectConfig } from '@push-based/user-flow-cli-testing';
 import { DEFAULT_RC_NAME } from '@push-based/user-flow';
-import { REMOTE_RC_JSON, REMOTE_USERFLOW_PATH } from '../rc-files/remote';
-import { REMOTE_USERFLOW_CONTENT } from '../user-flows/remote.uf';
-import { INITIATED_PRJ_BIN, INITIATED_PRJ_ROOT } from './initiated';
+import { UserFlowProjectConfig } from '@push-based/user-flow-cli-testing';
+import { INITIATED_PRJ_BIN, INITIATED_PRJ_ROOT } from '../initialized/cfg';
+import { REMOTE_RC_JSON, REMOTE_USERFLOW_PATH } from './rc';
+import { REMOTE_USERFLOW_CONTENT } from './flow1.uf';
 
 export const REMOTE_PRJ_CFG: UserFlowProjectConfig = {
   root: INITIATED_PRJ_ROOT,
