+++ packages/test-data/src/lib/sandboxes/remote/index.ts
@@ -1,0 +1,3 @@
+export * from './rc';
+export * from './cfg';
+export * from './flow1.uf';
