+++ packages/test-data/src/lib/sandboxes/remote/rc.ts
@@ -1,7 +1,7 @@
+import * as path from 'path';
 import { RcJson } from '@push-based/user-flow';
 import { SANDBOX_BASE_RC_JSON } from '@push-based/user-flow-cli-testing';
-import { REMOTE_USERFLOW_NAME, REMOTE_USERFLOW_TITLE } from '../user-flows/remote.uf';
-import * as path from 'path';
+import { REMOTE_USERFLOW_NAME, REMOTE_USERFLOW_TITLE } from './flow1.uf';
 
 export const REMOTE_RC_NAME = '.user-flow.remote.json';
 export const REMOTE_RC_JSON: RcJson = {
