+++ packages/test-data/src/lib/sandboxes/static/index.ts
@@ -1,0 +1,3 @@
+export * from './rc';
+export * from './cfg';
+export * from './flow1.uf';
