+++ packages/test-data/src/lib/sandboxes/static/rc.ts
@@ -1,6 +1,6 @@
 import { RcJson } from '@push-based/user-flow';
 import { SANDBOX_BASE_RC_JSON, SERVE_COMMAND_PORT } from '@push-based/user-flow-cli-testing';
-import { STATIC_USERFLOW_NAME, STATIC_USERFLOW_TITLE } from '../user-flows/static.uf';
+import { STATIC_USERFLOW_NAME, STATIC_USERFLOW_TITLE } from './flow1.uf';
 import * as path from 'path';
 
 export const STATIC_RC_NAME = '.user-flow.static.json';
