+++ packages/test-data/src/lib/sandboxes/static/static-app/favico.ts
