+++ packages/test-data/src/lib/sandboxes/static/static-app/index.html.ts
