+++ packages/test-data/src/lib/sandboxes/static/static-app/index.ts
@@ -1,0 +1,3 @@
+export * from './index.html';
+export * from './user-flow-square';
+export * from './favico';
