+++ packages/test-data/src/lib/user-flows/index.ts
@@ -1,0 +1,3 @@
+export * from './wrong-mod-export.example.uf';
+export * from './wrong-ext.example.uf';
+export * from './valide.example.uf';
